fiks antall dager

+++ formidling/src/main/java/no/nav/ung/sak/formidling/template/dto/innvilgelse/InnvilgelseDto.java
@@ -13,7 +13,7 @@ public record InnvilgelseDto(
     FellesDto felles,
     ResultatFlaggDto resultat,
     LocalDate ytelseFom,
-    int antallDager,
+    long antallDager,
     Set<TilkjentPeriodeDto> tilkjentePerioder,
     Set<GbeløpPeriodeDto> gbeløpPerioder,
     SatserDto satser) implements TemplateData {}
